refactor(web): move Landing to screens in favour of readability
