+++ packages/web/src/pages/index.tsx
@@ -16,6 +16,6 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import Landing from '../screens/landing/Landing';
+import Landing from '../screens/Landing.js';
 
 export default Landing;
